test: add tests for koin di modules

This commit renames `AppModules.kt` to `ApplicationModule.kt` and
refactors the module definitions to use the new Koin annotations for
better clarity and structure. Additionally, new test files are added to
verify the functionality of the application and core modules.

+++ feature/charts/src/main/kotlin/io/github/stslex/workeeper/feature/charts/domain/model/SingleChartDomainModel.kt
@@ -2,5 +2,5 @@ package io.github.stslex.workeeper.feature.charts.domain.model
 
 internal data class SingleChartDomainModel(
     val name: String,
-    val values: List<Double>,
+    val values: List<Float>,
 )
