+++ feature/charts/src/main/kotlin/io/github/stslex/workeeper/feature/charts/ui/mvi/model/SingleChartUiModel.kt
@@ -6,5 +6,5 @@ import kotlinx.collections.immutable.ImmutableList
 @Stable
 internal data class SingleChartUiModel(
     val name: String,
-    val properties: ImmutableList<Double>,
+    val properties: ImmutableList<Float>,
 )
